pu4c.nprandom && rename

+++ pcseg/data/dataset/semantickitti/semantickitti_cylinder.py
@@ -12,7 +12,7 @@ from torchsparse.utils.quantize import sparse_quantize
 
 from .semantickitti import SemantickittiDataset
 
-from tools.utils.common.seg_utils import aug_points
+from pcseg.utils.seg_utils import aug_points
 
 
 # transformation between Cartesian coordinates and polar coordinates
@@ -150,7 +150,7 @@ class SemkittiCylinderDataset(data.Dataset):
         cur_grid_size = self.grid_size
         intervals = crop_range / (cur_grid_size - 1)
 
-        point_coord = (np.floor((np.clip(xyz_pol, min_bound, max_bound) - min_bound) / intervals)).astype(np.int)
+        point_coord = (np.floor((np.clip(xyz_pol, min_bound, max_bound) - min_bound) / intervals)).astype(np.int32)
         voxel_coord, voxel_label, inds, inverse_map = voxelize_with_label(
             point_coord, point_label, len(self.class_names))
         voxel_centers = (voxel_coord.astype(np.float32) + 0.5) * intervals + min_bound
@@ -163,11 +163,11 @@ class SemkittiCylinderDataset(data.Dataset):
             'name': pc_data['path'],
             'point_feature': point_feature.astype(np.float32),
             'point_coord': point_coord.astype(np.float32),
-            'point_label': point_label.astype(np.int),
+            'point_label': point_label.astype(np.int32),
             'voxel_feature': voxel_feature.astype(np.float32),
-            'voxel_coord': voxel_coord.astype(np.int),
-            'voxel_label': voxel_label.astype(np.int),
-            'inverse_map': inverse_map.astype(np.int),
+            'voxel_coord': voxel_coord.astype(np.int32),
+            'voxel_label': voxel_label.astype(np.int32),
+            'inverse_map': inverse_map.astype(np.int32),
             'num_points': np.array([num_points_current_frame]),
         })
